support client 6.0.14.2
Send the "Seed" packet from 6.0.14.2 on, and use a different 0xb9
packet size.

+++ src/connection.h
@@ -43,7 +43,7 @@ struct stateful_client {
     struct uo_fragment_character_info characters[MAX_CHARACTERS];
     unsigned num_characters;
 
-    uint16_t supported_features_flags;
+    uint32_t supported_features_flags;
 
     unsigned char ping_request, ping_ack;
